+++ packages/tauri/src/virtual_branches/mod.rs
@@ -2,6 +2,9 @@ pub mod branch;
 pub use branch::Branch;
 pub mod target;
 
+mod files;
+pub use files::*;
+
 mod integration;
 pub use integration::GITBUTLER_INTEGRATION_BRANCH_NAME;
 
@@ -18,5 +21,8 @@ pub use iterator::BranchIterator as Iterator;
 #[cfg(test)]
 mod tests;
 
-mod vbranch;
-pub use vbranch::*;
+mod r#virtual;
+pub use r#virtual::*;
+
+mod remote;
+pub use remote::*;
